Model correction

+++ Backend_POS/Models/DTO/ZaglavljeRacuna/UpdateZaglavljeRacunaRequestDTO.cs
@@ -5,6 +5,6 @@
         public int Broj { get; set; }
         public DateTime Datum { get; set; } = DateTime.Now;
         public string Napomena { get; set; } = string.Empty;
-        public long KupacId { get; set; }
+        public int KupacId { get; set; }
     }
 }
